[rpi01] arm: Workaround potential hang in cpu_v6_do_idle
Entering WFI state with an outstanding instruction cache read in flight
can hang the ARM core, so you have to be very careful to preload the
region following the WFI into the cache before going to sleep. We do
this by going round a loop once with the mcr instructions conditionally
disabled before doing it for real.

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
[SA: added explanation from https://github.com/raspberrypi/linux/issues/20]
Signed-off-by: Simon Arlott <simon@fire.lp0.eu>