MIPS: bcm63xx: add support for the BCM63168 CPU
This CPU has a 5 digit model, so the CPU ID has been increased to u32
and 5 digit IDs are checked before 4 digit IDs.

Some flash initialisation and reconfiguration is required to be able
to read the CFE boot configuration.

The IRQ controller supports 128 IRQs, so NR_IRQs is increased to
accommodate this.

The same registers and IRQs should work for the BCM63268 but I have
no hardware to test with.

Signed-off-by: Simon Arlott <simon@fire.lp0.eu>