[PATCH linux-next v2 2/2] MIPS: bmips: Add bcm6345-l1 interrupt controller
Add the BCM6345 interrupt controller based on the SMP-capable BCM7038
and the BCM3380 but with packed interrupt registers.

Add the BCM6345 interrupt controller to a list with the existing BCM7038
so that interrupts on CPU1 are not ignored.

Update the maintainers file list for BMIPS to include this driver.

Signed-off-by: Simon Arlott <simon@fire.lp0.eu>
---
v2: Removed #ifdef CONFIG_SMP as the effect on UP is a single direct
    array access to discover that __cpu_logical_map[0] == 0.

    Added a comment regarding the locking.

    Renamed "idx" to "i" or "cpu_idx" as appropriate.

    Removed #ifdef __BIG_ENDIAN as all the BCM6345-dervied MIPS SoCs
    are BE.

v1: Rebased against linux-next-20151209, no other changes.