leds-bcm6328: Reuse bcm6328_led_set() instead of copying its functionality
When ensuring a consistent initial LED state in bcm6328_led (as they may
be blinking instead of on/off), the LED register is set using a copy of
bcm6328_led_set(). To avoid further errors relating to active low handling,
call this function directly instead.

As bcm6328_led_set() expects to acquire the spinlock, narrow the locking
to only cover reading of the current state. There is no need to hold the
spinlock between reading the current value and setting it again because
the LED device has not yet been registered.

Signed-off-by: Simon Arlott <simon@fire.lp0.eu>