dmaengine: virt-dma: allow the completion tasklet to be deferred
When the dmaengine has to run a tasklet itself (because it is handling
individual network packet transfers), it needs to be able to dispatch
received packets and check for further completed descriptors before
re-enabling interrupts.

Instead of trying to schedule multiple tasklets which would need to be
strictly ordered to achieve the correct result, add a variant of
vchan_cookie_complete() that does not schedule the completion tasklet.

Since the completion tasklet will now not be run automatically, add a
function vchan_complete_task() to call the process directly.

Signed-off-by: Simon Arlott <simon@fire.lp0.eu>